requirements.txt
flask 
pyjwt 
python-dotenv
sqlalchemy
pymysql
cryptography
python-dotenv
